Dissolve SexpNode into Node.
